fix: 修复 RecommendationService 编译类型错误
tagWeightMap.getOrDefault 返回 Integer，需要显式转为 double

+++ backend/src/main/java/com/learning/platform/service/RecommendationService.java
@@ -82,7 +82,7 @@ public class RecommendationService {
                         new QueryWrapper<ResourceTag>().eq("tag_id", tagId));
                 for (ResourceTag rt : rts) {
                     if (!interactedResourceIds.contains(rt.getResourceId())) {
-                        tagScores.merge(rt.getResourceId(), tagWeightMap.getOrDefault(tagId, 1.0), Double::sum);
+                        tagScores.merge(rt.getResourceId(), tagWeightMap.getOrDefault(tagId, 1).doubleValue(), Double::sum);
                     }
                 }
             }
